Converted form component to typescript
