--- packages/patternfly-4/react-core/src/components/Form/ActionGroup.d.ts
@@ -1,7 +1,0 @@
-import { FunctionComponent, HTMLProps } from 'react';
-
-export interface ActionGroupProps extends HTMLProps<HTMLDivElement> { }
-
-declare const ActionGroup: FunctionComponent<ActionGroupProps>;
-
-export default ActionGroup;
