--- packages/patternfly-4/react-core/src/components/Form/FormContext.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const FormContext = React.createContext({
-  isHorizontal: false
-});
