+++ packages/patternfly-4/react-core/src/components/Form/FormContext.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const FormContext = React.createContext({
+  isHorizontal: false
+});
