--- packages/patternfly-4/react-core/src/components/Form/FormGroup.d.ts
@@ -1,16 +1,0 @@
-import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { Omit } from '../../helpers/typeUtils'
-
-export interface FormGroupProps extends Omit<HTMLProps<HTMLDivElement>, 'label'> {
-  isValid?: boolean;
-  isRequired?: boolean;
-  label?: ReactNode;
-  helperText?: ReactNode;
-  helperTextInvalid?: ReactNode;
-  fieldId: string;
-  isInline?: boolean;
-}
-
-declare const FormGroup: FunctionComponent<FormGroupProps>;
-
-export default FormGroup;
