--- packages/patternfly-4/react-core/src/components/Form/FormHelperText.d.ts
@@ -1,12 +1,0 @@
-import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-
-export interface FormHelperTextProps extends HTMLProps<HTMLDivElement> {
-  children?: ReactNode;
-  isError?: boolean;
-  isActive?: boolean;
-  className?: string;
-}
-
-declare const FormHelperText: FunctionComponent<FormHelperTextProps>;
-
-export default FormHelperText;
