--- packages/patternfly-4/react-core/src/components/Form/index.d.ts
@@ -1,4 +1,0 @@
-export { default as ActionGroup, ActionGroupProps } from './ActionGroup';
-export { default as Form, FormProps } from './Form';
-export { default as FormGroup, FormGroupProps } from './FormGroup';
-export { default as FormHelperText } from './FormHelperText';
