--- packages/patternfly-4/react-core/src/components/Form/index.js
@@ -1,4 +1,0 @@
-export { default as ActionGroup } from './ActionGroup';
-export { default as Form } from './Form';
-export { default as FormGroup } from './FormGroup';
-export { default as FormHelperText } from './FormHelperText';
